# docs/swagger.yaml
definitions:
  admin.ListProjectsResponse:
    properties:
      data:
        $ref: '#/definitions/admin.ListProjectsResponseData'
      error_msg:
        type: string
    type: object
  admin.ListProjectsResponseData:
    properties:
      results:
        items:
          $ref: '#/definitions/admin.ListProjectsResponseDataResult'
        type: array
      total:
        type: integer
    type: object
  admin.ListProjectsResponseDataResult:
    properties:
      description:
        type: string
      id:
        type: string
      nickname:
        type: string
      tags:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
info:
  contact: {}
  title: Linux Do CDK
  version: 0.1.0
paths:
  /api/v1/admin/projects/report:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        - 2
        in: query
        name: status
        type: integer
        x-enum-varnames:
        - ProjectStatusNormal
        - ProjectStatusHidden
        - ProjectStatusViolation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ListProjectsResponse'
      tags:
      - admin
...
swagger: "2.0"